Perform CRC validation of firmware in flash before switching to it
Using CRC HAL for now. Seems a bit bulky for the task but it was the
fastest way to get values matching zlib CRC-32 by referencing other
people using the HAL

+++ firmware-hc/crc.h
@@ -5,6 +5,6 @@
 
 void crc_init();
 
-u32 crc_32(const u32 *din, int count);
+u32 crc_32(const u8 *din, int count);
 
 #endif
